refactor: Names 생성 및 관련 메서드들 이동
